refactor: 디테일 페이지 레이아웃 추가
