+++ src/components/layouts/DetailPageLayout.vue
